requirements.txt
pymongo==2.6.3
numpy==1.8.2
matplotlib==1.5.1
kafka-python==1.3.2
